requirements.txt
Flask==2.1.2
Flask-Classful==0.14.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
werkzeug==2.1.2
openai==0.28.0
requests==2.31.0
azure-storage-blob==12.16.0